libebml2: move headers to ebml2
so it doesn't collide/mismatch with libebml
