+++ libebml2/ebml2/ebml.h
@@ -31,7 +31,7 @@
 
 #include "corec/node/node.h"
 #include "corec/helpers/parser/parser.h"
-#include "ebml/ebml_config.h"
+#include "ebml2/ebml_config.h"
 
 #ifdef __cplusplus
 extern "C" {
