+++ libmatroska2/matroska/matroska.h
@@ -28,7 +28,7 @@
 #ifndef MATROSKA_MATROSKA_H
 #define MATROSKA_MATROSKA_H
 
-#include "ebml/ebml.h"
+#include "ebml2/ebml.h"
 #include "matroska/matroska_config.h"
 
 #if defined(matroska2_EXPORTS)
